+++ app/controllers/api/verify/password_confirm_controller.rb
@@ -1,6 +1,6 @@
 module Api
   module Verify
-    class PasswordConfirmController < Api::BaseController
+    class PasswordConfirmController < BaseController
       def create
         result, personal_key = Api::ProfileCreationForm.new(
           password: verify_params[:password],
@@ -21,6 +21,10 @@ module Api
       end
 
       private
+
+      def required_step
+        'password_confirm'
+      end
 
       def verify_params
         params.permit(:password, :user_bundle_token)
